requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
Flask-Limiter
idna
itsdangerous
Jinja2==3.1.6
limits
markdown-it-py
MarkupSafe
pillow
rembg